chore: move toKebabCase helper
