+++ packages/react-ui/internal/CommonWrapper/__tests__/toKebabCase.test.ts
@@ -1,4 +1,4 @@
-import { toKebabCase } from '../toKebabCase';
+import { toKebabCase } from '../../../lib/toKebabCase';
 
 it('convert to kebab case', () => {
   expect(toKebabCase('toKebabCase')).toBe('to-kebab-case');
